pyproject.toml
[project]
name = "dragster_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "dagster",
    "dagster-cloud",
    "pymongo",
    "redis",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver",
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dragster_project.definitions"
code_location_name = "dragster_project"

[tool.setuptools.packages.find]
exclude=["dragster_project_tests"]
